requirements.txt
fastapi
uvicorn[standard]
pydantic<2
SQLAlchemy[asyncio]
asyncpg
alembic
redis
passlib[bcrypt]
python-jose[cryptography]
pytest
pytest-cov
httpx
bcrypt<4
motor
pymongo
python-logstash
sentry-sdk==0.19.5
opentelemetry-api==1.17.0
opentelemetry-sdk==1.17.0
opentelemetry-instrumentation-fastapi==0.38b0
opentelemetry-exporter-jaeger==1.17.0
opentelemetry-exporter-otlp==1.17.0
opentelemetry-exporter-otlp-proto-http
requests